Initial OSX DriveInfo changes
