+++ src/Common/src/Interop/OSX/Interop.Libraries.cs
@@ -5,8 +5,10 @@ internal static partial class Interop
 {
     private static partial class Libraries
     {
+        internal const string INODE64SUFFIX = "$INODE64";
         internal const string CoreFoundationLibrary = "/System/Library/Frameworks/CoreFoundation.framework/CoreFoundation";
         internal const string CoreServicesLibrary   = "/System/Library/Frameworks/CoreServices.framework/CoreServices";
+        internal const string libc = "libc";
         internal const string libproc = "libproc";
         internal const string LibSystemKernel = "/usr/lib/system/libsystem_kernel";
     }
